Module convey_rs::property

Structs

Argument

ArgumentHistory

ArrayProperty

ArrayPropertyStruct

BaseHistory

ByteProperty

Color

FINGPUT1Buffer

FINGPUT1BufferPixel

FINLuaProcessorStateStorage

FINLuaProcessorStateStorageStruct

FINNetworkTrace

FrameRange

InventoryItem

InventoryStack

ItemAmount

MapProperty

NoneHistory

Property

SetProperty

StringTableEntryHistory

StructPropertyBox

StructPropertyInventoryItem

StructPropertyRailroadTrackPosition

TextProperty

TransformHistory

Enums

ArrayPropertyStructValue

ArrayPropertyValue

FINLuaProcessorStateStorageStructValue

MapPropertyKey

MapPropertyValue

Numeric

A container is needed for these types because many fields are “upgraded” from floats to doubles with new file versions, and other times types may contain multiple inner types

PropertyValue

SetPropertyValue

StructPropertyValue

TextPropertyHistory